Begin refactoring
