+++ go.mod
@@ -1,5 +1,5 @@
 module github.com/ondrejmalina/json-parser
 
-go 1.22.4
+go 1.24.0
 
-require golang.org/x/text v0.20.0
+require golang.org/x/text v0.24.0
