+++ internal/lexer/lexer_test.go
